MDEV-27169 Change default values of Spider server variables to default values of table variables
As the first step to deprecate Spider table options, we need to change
the default values of the Spider server variables from -1 (use table value)
to the default table values.

For example, the current default server value of spider_auto_increment_mode
is -1 (use table value) and the current default table value is 0 (normal mode).
Thus, the real default value is 0. we do like to remove this indirection
(just set the default server value to 0).

The default table values are defined in spider_set_connect_info_default().